tests: root_test: move btrfs test to new file
